==> 1_SDMbench_download.ipynb <==
# Download h5ad

from concurrent.futures import ProcessPoolExecutor
from pathlib import Path

import pandas as pd
import requests

metadata_file = "downloads.tsv"

download_path = Path("./data")

h5ad_path = download_path / "sdmbench"

metadata = pd.read_table(metadata_file, index_col=0)

def download_file(url: str):
    filename = h5ad_path / url.split("/")[-1]
    with requests.get(url, stream=True) as response:
        response.raise_for_status()
        with open(filename, "wb") as file:
            for chunk in response.iter_content(chunk_size=1024):
                file.write(chunk)
    return filename

with ProcessPoolExecutor(max_workers=16) as pool:
    downloaded_files = list(pool.map(download_file, metadata["url"].to_list()))

# Download stats

import urllib.request

url = "https://static-content.springer.com/esm/art%3A10.1038%2Fs41592-024-02215-8/MediaObjects/41592_2024_2215_MOESM6_ESM.xlsx"

urllib.request.urlretrieve(url, download_path / "sdmbench_metrics.xlsx")
