==> 2_BayesSpace_DLPFC.ipynb <==
suppressPackageStartupMessages(library(BayesSpace))

data_dir <- "./data/LIBD_DLPFC"
result_dir <- "./results/LIBD_DLPFC"

set.seed(42)

get_SingleCellExperiment <- function(path) {
    feature_file <- file.path(path, "features.tsv")
    observation_file <- file.path(path, "observations.tsv")
    matrix_file <- file.path(path, "counts.mtx")

    rowData <- read.delim(feature_file, stringsAsFactors = FALSE, row.names = 1)
    colData <- read.delim(observation_file, stringsAsFactors = FALSE, row.names = 1)
    counts <- as(Matrix::t(Matrix::readMM(matrix_file)), "CsparseMatrix")

    sce <- SingleCellExperiment(
        rowData = rowData,
        colData = colData,
        assays = list(counts = counts)
    )
    return(sce)
}

run_BayesSpace <- function(
    sce,
    n_clusters = 7,
    platform = "Visium",
    n_components = 15,
    n_HVG = 2000) {
    
    sce <- spatialPreprocess(
        sce,
        platform = platform,
        n.PCs = n_components,
        n.HVGs = n_HVG,
        log.normalize = TRUE
    )

    sce <- spatialCluster(
        sce,
        q = n_clusters,
        d = n_components,
        platform = platform
    )
    return(sce)
}

write_labels_to_tsv <- function(sce, path, col_name = "spatial.cluster") {
    label_df <- as.data.frame(colData(sce))[c(col_name)]

    dir.create(path, showWarnings = FALSE, recursive = TRUE)

    colnames(label_df) <- c("label")
    write.table(
        label_df,
        file = file.path(path, "BayesSpace.tsv"),
        sep = "\t",
        col.names = NA,
        quote = FALSE
    )
}

metadata <- read.delim(
    file.path(data_dir, "samples.tsv"),
    stringsAsFactors = FALSE,
    row.names = 1
)[, c("directory", "n_clusters")]

for (i in 1:nrow(metadata)) {
    sample <- metadata[i, ]
    print(paste("Processing", sample[["directory"]]))
    sce <- get_SingleCellExperiment(file.path(data_dir, sample[["directory"]]))
    sce <- run_BayesSpace(sce, n_clusters = sample[["n_clusters"]])
    write_labels_to_tsv(sce, file.path(result_dir, sample[["directory"]]))
}
