==> 2_SpiceMix.ipynb <==
import os
from pathlib import Path

import pandas as pd

from utils import get_anndata

data_dir = Path("./data/LIBD_DLPFC")
result_dir = Path("./results/SpiceMix")

seed = 42

result_dir.mkdir(parents=True, exist_ok=True)

metadata = pd.read_table(
    data_dir / "samples.tsv", usecols=["directory", "n_clusters"]
).set_index("directory")

n_genes = 3_000
n_pcs = 30
k = 20

sample = metadata.loc["Br8100_151673", :]

# generate the input file and preprocess
import scanpy as sc
from popari.components import PopariDataset
from popari.io import save_anndata

adata = get_anndata(data_dir / sample.name)

sc.pp.normalize_total(adata, target_sum=1e4)
sc.pp.log1p(adata)
sc.pp.highly_variable_genes(adata, n_top_genes=n_genes)

adata = PopariDataset(adata[:, adata.var["highly_variable"]], "processed")
adata.compute_spatial_neighbors()

save_anndata(result_dir / f"{sample.name}.h5ad", [adata])

conda_path = "~/miniconda3/bin/activate"

log_dir = result_dir / "logs"
log_dir.mkdir(parents=True, exist_ok=True)


path = Path(os.getcwd())

sample_log_dir = log_dir.resolve() / sample.name
out_dir = result_dir / sample.name

sample_log_dir.mkdir(parents=True, exist_ok=True)
out_dir.mkdir(parents=True, exist_ok=True)

conda_env = "spicemix"
conda_cmd = f"source {conda_path} {conda_env}"

for reg in [12, 10, 8, 6, 4]:
    cmd = (
        f"{path/'spicemix.py'} "
        f"-i {result_dir / f'{sample.name}.h5ad'} "
        f"-o {out_dir/f'spicemix_exp{reg}.h5ad'} "
        f"--reg_exp {reg} "
        f"--k {k} "
        f"--seed {seed} "
    )

    os.popen(
        f"sbatch -J {sample.name} --mem=64G -n 8 -N 1 --gres=gpu:1 --partition graphical-legacy --time=06:00:00 "
        f"-o {sample_log_dir/f'spicemix_{reg}'}.txt "
        f'--wrap="{conda_cmd} && {cmd}" '
    ).read()
